Фиксы

+++ PolygonalImageSegmentation/src/main/java/controllers/AlgorithmMenuLayoutController.java
@@ -145,22 +145,22 @@ public class AlgorithmMenuLayoutController implements Observer {
 
     @FXML
     public void handleApplyHough(ActionEvent event) {
-        int distance = (int) spinnerHoughDistance.getValue();
-        int angle = (int) spinnerHoughAngle.getValue();
+        double distance = (double) spinnerHoughDistance.getValue();
+        double angle = (double) spinnerHoughAngle.getValue();
         int threshold = (int) thresholdHoughSlider.getValue();
 
         RadioButton button = (RadioButton) toggleGroupHough.getSelectedToggle();
         if (button.getId().equals("radiobHoughClassic")) {
-            int srn = Integer.parseInt(srnTextField.getText());
-            int stn = Integer.parseInt(stnTextField.getText());
-            int minTheta = Integer.parseInt(minThetaTextField.getText());
-            int maxTheta = Integer.parseInt(maxThetaTextField.getText());
+            double srn = Double.parseDouble(srnTextField.getText());
+            double stn = Integer.parseInt(stnTextField.getText());
+            double minTheta = Integer.parseInt(minThetaTextField.getText());
+            double maxTheta = Integer.parseInt(maxThetaTextField.getText());
             imagesHandler.doHoughConversion(true, distance, angle, threshold, srn, stn, minTheta, maxTheta);
             return;
         }
         if (button.getId().equals("radiobHoughProbably")) {
-            int maxLineGap = (int) spinnerHoughProbablyMaxLineGap.getValue();
-            int minLineLength = (int) spinnerHoughProbablyMinLineLength.getValue();
+            double maxLineGap = (double) spinnerHoughProbablyMaxLineGap.getValue();
+            double minLineLength = (double) spinnerHoughProbablyMinLineLength.getValue();
             imagesHandler.doHoughConversion(false, distance, angle, threshold, maxLineGap, minLineLength);
         }
     }
